10/04/2026/(3)

+++ src/main/java/asm/poly/asm_java6/repository/BrandRepository.java
@@ -3,6 +3,5 @@ package asm.poly.asm_java6.repository;
 import asm.poly.asm_java6.enity.Brand;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BrandRepository extends JpaRepository<Brand, Integer> {
-
+public interface BrandRepository extends JpaRepository<Brand, Long> {
 }
